Moved PanacheQuery to an impl class, added an interface in its place
So I don't have to document the public constructor
